% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{brand.colors}
\alias{brand.colors}
\title{MBLabs color ramp}
\usage{
brand.colors(
  x = length(pal()),
  colors = pal()[!names(pal()) \%in\% c("white", "black", "light", "gray")],
  alpha = 0.9,
  ...
)
}
\arguments{
\item{x}{number of colors to interpolate}

\item{colors}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}

\item{alpha}{logical: should alpha channel (opacity) values be
    returned?   It is an error to give a true value if
    \code{space} is specified.}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRamp]{grDevices::colorRamp}}
  \describe{
    \item{\code{bias}}{a positive number.  Higher values give more widely spaced
    colors at the high end.}
    \item{\code{space}}{a character string; interpolation in RGB or \abbr{CIE} Lab
    color spaces.}
    \item{\code{interpolate}}{use spline or linear interpolation.}
  }}
}
\value{
A function to interpolate colors
}
\description{
Qualitative color ramp derived from active branding.
}
\examples{
x <- rchisq(100, df=4)
hist(x, freq=FALSE, ylim=c(0, 0.2), col=brand.colors(20), border="white")
hist(x, freq=FALSE, ylim=c(0, 0.2), col=brand.colors(8, alpha=0.5), border="white")

}
