% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Modified base R plot with Bootstrap branding}
\usage{
plot(
  x,
  y = NULL,
  axes = c("bottomright", "topright", "bottomleft", "topleft"),
  bg = NULL,
  col = pal(1),
  lwd = 2,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  nx = NA,
  ny = NULL,
  grid.lty = "dotted",
  grid.lwd = 2,
  ...
)
}
\arguments{
\item{x, y}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}

\item{axes}{position of X (\code{bottom} or \code{top}) and Y (\code{left} or \code{right}) axes (default: \code{bottomright})}

\item{main}{a main title for the plot, see also \code{\link[graphics]{title}}.}

\item{sub}{a subtitle for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{a label for the y axis, defaults to a description of \code{y}.}

\item{...}{
  Arguments passed on to \code{\link[graphics:plot.default]{graphics::plot.default}}
  \describe{
    \item{\code{x,y}}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}
    \item{\code{type}}{1-character string giving the type of plot desired.  The
    following values are possible, for details, see \code{\link[base]{plot}}:
    \code{"p"} for points,
    \code{"l"} for lines,
    \code{"b"} for both points and lines,
    \code{"c"} for empty points joined by lines,
    \code{"o"} for overplotted points and lines,
    \code{"s"} and \code{"S"} for stair steps and
    \code{"h"} for histogram-like vertical lines.  Finally,
    \code{"n"} does not produce any points or lines.}
    \item{\code{xlim}}{the x limits (x1, x2) of the plot.  Note that \code{x1 > x2}
    is allowed and leads to a \sQuote{reversed axis}.

    The default value, \code{NULL}, indicates that the range of the
    \link{finite} values to be plotted should be used.}
    \item{\code{ylim}}{the y limits of the plot.}
    \item{\code{log}}{a character string which contains \code{"x"} if the x axis
    is to be logarithmic, \code{"y"} if the y axis is to be logarithmic
    and \code{"xy"} or \code{"yx"} if both axes are to be logarithmic.}
    \item{\code{ann}}{a logical value indicating whether the default annotation
    (title and x and y axis labels) should appear on the plot.}
    \item{\code{frame.plot}}{a logical indicating whether a box should be drawn
    around the plot.}
    \item{\code{panel.first}}{an \sQuote{expression} to be evaluated after the
    plot axes are set up but before any plotting takes place.  This can
    be useful for drawing background grids or scatterplot smooths.  Note
    that this works by lazy evaluation: passing this argument from other
    \code{plot} methods may well not work since it may be evaluated too
    early.}
    \item{\code{panel.last}}{an expression to be evaluated after plotting has
    taken place but before the axes, title and box are added.  See the
    comments about \code{panel.first}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio, see \code{\link[graphics]{plot.window}}.}
    \item{\code{xgap.axis,ygap.axis}}{the \eqn{x/y} axis gap factors, passed as
    \code{gap.axis} to the two \code{\link[graphics]{axis}()} calls (when
    \code{axes} is true, as per default).}
  }}
}
\value{
base R plot
}
\description{
Shorthand function to generate base R plots with Mel B. Labs branding. Many tips derived from \href{https://r-charts.com/base-r/margins/}{r-charts.com}, since base R graphic documentation is often lacking.
}
\examples{
set.seed(1)
x <- runif(100, min = -5, max = 5)
y <- x ^ 3 + rnorm(100, mean = 0, sd = 5)

plot(x, y, main="Mel B. Labs Branded Plot", sub="Scatter plot")
plot(x, y, nx=NULL,
  main="Mel B. Labs Branded Plot", sub="Subtitle",
  xlab="X Units", ylab="Y Units")
abline(h=0, col=pal("red"), lwd=2)

plot(x, type="h", col=pal(c("red", "green"))[(x > 0) + 1],
  main="Mel B. Labs Branded Plot", sub="Histogram")

par.labs("family")
par("fg")
palette()

}
