% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{brand}
\alias{brand}
\title{Fetch Bootstrap branding}
\usage{
brand(file = "_brand.yml", ...)
}
\arguments{
\item{file}{path to a structured \verb{_brand.yml} file}

\item{...}{
  Arguments passed on to \code{\link[yaml:read_yaml]{yaml::read_yaml}}
  \describe{
    \item{\code{fileEncoding}}{character string: if non-empty declares the
    encoding used on a file (not a connection) so the character data can
    be re-encoded.  See \code{\link{file}}.}
    \item{\code{text}}{character string: if \code{file} is not supplied and this is,
    then data are read from the value of \code{text} via a text connection.
    Notice that a literal string can be used to include (small) data sets
    within R code.}
    \item{\code{error.label}}{a label to prepend to error messages (see Details).}
    \item{\code{readLines.warn}}{logical (default:TRUE) suppress warnings from readLines used inside read_yaml}
  }}
}
\value{
A list of branding elements
}
\description{
Make this package compatible with Quarto \code{brand.yml} branding features. Branding can be read from a project file, or else it will use this package defaults.
}
\examples{
brand()$color$palette

}
\seealso{
\href{https://posit-dev.github.io/brand-yml/}{brand.yml}
}
