% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{brand}
\alias{brand}
\title{Fetch Bootstrap branding}
\usage{
brand(file = "_brand.yml")
}
\arguments{
\item{file}{path to \verb{_brand.yml} brand configuration file, normally this file is auto-detected in the working tree, but may be specified here to swap branding dynamically}
}
\value{
A list of branding elements
}
\description{
Make this package compatible with \code{brand.yml} unified branding features. Branding can be read from a configuration file, else it will use one of these package built-in themes. Refer to \href{https://posit-dev.github.io/brand-yml/}{brand.yml} documentation.
}
\examples{
scales::show_col(unlist(brand()$color$palette))
.globals$brand$meta$name

}
\references{
\href{https://posit-dev.github.io/brand-yml/}{brand.yml}
}
