% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gglabs}
\alias{gglabs}
\title{Modified ggplot2 objects for Mel B. Labs blog posts}
\usage{
gglabs(
  data = NULL,
  mapping = aes(),
  pos_x = c("bottom", "top"),
  pos_y = c("right", "left"),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{pos_x}{Position of x-axis (bottom or top)}

\item{pos_y}{Position of y-axis (right or left)}

\item{...}{
  Arguments passed on to \code{\link[=theme_labs]{theme_labs}}
  \describe{
    \item{\code{base_bg}}{Plot, panel, legend background}
    \item{\code{base_color}}{Color for text and line elements}
    \item{\code{base_size}}{base font size, given in pts.}
    \item{\code{base_family}}{base font family}
  }}
}
\value{
A \code{ggplot2} object with new theme elements applied
}
\description{
Convenience function to generate \code{ggplot2} plots with custom element sizes,
color scales and guides matching active branding (per \verb{_brand.yml}).
}
\examples{
require(ggplot2)

gglabs(mtcars, aes(factor(carb), mpg, fill=factor(carb))) + 
  geom_col() +
  guides(y=guide_none(), y.sec=guide_axis()) +
  labs(
    x = "carb",
    title = "My Long and Descriptive Plot Title",
    subtitle = "My very long subtitle with many units",
    caption = "My very long plot caption with many references.")

gglabs(mtcars, aes(wt, mpg)) +
  geom_smooth(color=pal("red"), fill=pal("pink")) +
  geom_point(size=3, aes(color=carb)) +
  labs(
  title = "My Title",
  subtitle = "My Subtitle",
  caption = "My plot caption")

gglabs(mtcars, aes(wt, mpg, color=factor(cyl))) +
  geom_point(size=3) + 
  labs(
  x = "cyl",
  title = "My Title",
  subtitle = "My Subtitle",
  caption = "My plot caption")

}
