% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{pal}
\alias{pal}
\title{Brand color palette}
\usage{
pal(x, named = TRUE)
}
\arguments{
\item{x}{color index or name(s), if missing returns the entire color palette}

\item{named}{return color names (default: TRUE)}
}
\value{
A vector of (named) hex color codes extracted from Bootstrap branding
}
\description{
Utility function to return a color palette derived from the active Bootstrap theme and branding elements. By default colors are read from a stuctured \verb{_brand.yaml} file (or from this package default brand).
}
\examples{
pal = pal()
pie(rep(1, length(pal)), col=pal, labels=names(pal), 
border="white", clockwise=TRUE)

}
\seealso{
\href{https://posit-dev.github.io/brand-yml/}{brand.yml}
}
